+++ Pronto-MIA/Tests/TestDataAccess/TestManagers/TestUserManager.cs
@@ -1,24 +1,30 @@
 namespace Tests.TestDataAccess.TestManagers
 {
+    using System.Linq;
+    using System.Text;
     using System.Threading.Tasks;
     using HotChocolate.Execution;
+    using Microsoft.EntityFrameworkCore;
     using Microsoft.Extensions.Logging;
     using NSubstitute;
+    using Pronto_MIA.BusinessLogic.Security;
     using Pronto_MIA.DataAccess;
     using Pronto_MIA.DataAccess.Managers;
+    using Pronto_MIA.Domain.Entities;
     using Xunit;
 
     public class TestUserManager
     {
         private readonly UserManager userManager;
+        private readonly ProntoMiaDbContext dbContext;
 
         public TestUserManager()
         {
-            var dbContext = TestHelpers.InMemoryDbContext;
-            TestDataProvider.InsertTestData(dbContext);
+            this.dbContext = TestHelpers.InMemoryDbContext;
+            TestDataProvider.InsertTestData(this.dbContext);
 
             this.userManager = new UserManager(
-                dbContext,
+                this.dbContext,
                 TestHelpers.TestConfiguration,
                 Substitute.For<ILogger<UserManager>>());
         }
@@ -30,6 +36,30 @@ namespace Tests.TestDataAccess.TestManagers
                 await this.userManager.Authenticate("Bob", "Hello Alice");
 
             Assert.NotEmpty(result);
+        }
+
+        [Fact]
+        public async Task TestAuthenticateUserOldHash()
+        {
+            var generator = new NullGenerator(null);
+            var user = new User(
+                "OldHash",
+                generator.HashPassword("Password"),
+                generator.GetIdentifier(),
+                "{}");
+            this.dbContext.Users.Add(user);
+            await this.dbContext.SaveChangesAsync();
+
+            var result =
+                await this.userManager.Authenticate("OldHash", "Password");
+
+            var adjustedUser = await this.dbContext.Users.FindAsync(user.Id);
+            Assert.NotEmpty(result);
+            Assert.Equal(
+                Pbkdf2Generator.Identifier, adjustedUser.HashGenerator);
+
+            this.dbContext.Users.Remove(adjustedUser);
+            await this.dbContext.SaveChangesAsync();
         }
 
         [Fact]
@@ -64,6 +94,198 @@ namespace Tests.TestDataAccess.TestManagers
             var user = await this.userManager.GetByUserName("_Alice");
 
             Assert.Null(user);
+        }
+
+        [Fact]
+        public async Task TestGetAllMultiple()
+        {
+            var users = this.userManager.GetAll();
+
+            Assert.Equal(2, await users.CountAsync());
+        }
+
+        [Fact]
+        public async Task TestGetAllEmpty()
+        {
+            this.dbContext.Users.RemoveRange(
+                this.dbContext.Users);
+            await this.dbContext.SaveChangesAsync();
+
+            var deploymentPlans = this.userManager.GetAll();
+            Assert.Empty(deploymentPlans);
+
+            TestDataProvider.InsertTestData(this.dbContext);
+        }
+
+        [Fact]
+        public async Task TestCreate()
+        {
+            var user = await this.userManager.Create(
+                "Alice2", "H1-ello");
+
+            var dbUser = await this.dbContext.Users.FindAsync(user.Id);
+            Assert.NotNull(user);
+
+            this.dbContext.Users.Remove(user);
+            await this.dbContext.SaveChangesAsync();
+        }
+
+        [Fact]
+        public async Task TestCreateExisting()
+        {
+            var error = await Assert.ThrowsAsync<QueryException>(async () =>
+            {
+                await this.userManager.Create(
+                    "Alice", "H1-ello");
+            });
+
+            Assert.Equal(
+                Error.UserAlreadyExists.ToString(),
+                error.Errors[0].Code);
+        }
+
+        [Fact]
+        public async Task TestCreateNonConformPassword()
+        {
+            var error = await Assert.ThrowsAsync<QueryException>(async () =>
+            {
+                await this.userManager.Create(
+                    "Alice2", "H-ello");
+            });
+
+            Assert.Equal(
+                Error.PasswordTooWeak.ToString(),
+                error.Errors[0].Code);
+            var user = this.dbContext.Users.Where(u => u.UserName == "Alice2");
+            Assert.Empty(user);
+        }
+
+        [Fact]
+        public async Task TestUpdate()
+        {
+            var user = new User(
+                "Alice2", new byte[10], NullGenerator.Identifier, "{}");
+            this.dbContext.Add(user);
+            await this.dbContext.SaveChangesAsync();
+
+            await this.userManager.Update(user.Id, "Alice3", "H1-ello");
+
+            user = await this.dbContext.Users.FindAsync(user.Id);
+            Assert.Equal("Alice3", user.UserName);
+            Assert.Equal(
+                "H1-ello", Encoding.ASCII.GetString(user.PasswordHash));
+
+            this.dbContext.Remove(user);
+            await this.dbContext.SaveChangesAsync();
+        }
+
+        [Fact]
+        public async Task TestUpdateUsernameOnly()
+        {
+            var user = new User(
+                "Alice2", new byte[10], NullGenerator.Identifier, "{}");
+            this.dbContext.Add(user);
+            await this.dbContext.SaveChangesAsync();
+
+            await this.userManager.Update(user.Id, "Alice3", null);
+
+            user = await this.dbContext.Users.FindAsync(user.Id);
+            Assert.Equal("Alice3", user.UserName);
+            Assert.Equal(
+                new byte[10], user.PasswordHash);
+
+            this.dbContext.Remove(user);
+            await this.dbContext.SaveChangesAsync();
+        }
+
+        [Fact]
+        public async Task TestUpdatePasswordOnly()
+        {
+            var user = new User(
+                "Alice2", new byte[10], NullGenerator.Identifier, "{}");
+            this.dbContext.Add(user);
+            await this.dbContext.SaveChangesAsync();
+
+            await this.userManager.Update(user.Id, null, "H1-ello");
+
+            user = await this.dbContext.Users.FindAsync(user.Id);
+            Assert.Equal("Alice2", user.UserName);
+            Assert.Equal(
+                "H1-ello", Encoding.ASCII.GetString(user.PasswordHash));
+
+            this.dbContext.Remove(user);
+            await this.dbContext.SaveChangesAsync();
+        }
+
+        [Fact]
+        public async Task TestUpdateEmpty()
+        {
+            var user = new User(
+                "Alice2", new byte[10], NullGenerator.Identifier, "{}");
+            this.dbContext.Add(user);
+            await this.dbContext.SaveChangesAsync();
+
+            await this.userManager.Update(user.Id, null, null);
+
+            user = await this.dbContext.Users.FindAsync(user.Id);
+            Assert.Equal("Alice2", user.UserName);
+            Assert.Equal(
+                new byte[10], user.PasswordHash);
+
+            this.dbContext.Remove(user);
+            await this.dbContext.SaveChangesAsync();
+        }
+
+        [Fact]
+        public async Task TestUpdateToExistingUsername()
+        {
+            var user = new User(
+                "Alice2", new byte[10], NullGenerator.Identifier, "{}");
+            this.dbContext.Add(user);
+            await this.dbContext.SaveChangesAsync();
+
+            var error = await Assert.ThrowsAsync<QueryException>(async () =>
+            {
+                await this.userManager.Update(user.Id, "Alice", null);
+            });
+
+            Assert.Equal(
+                Error.UserAlreadyExists.ToString(),
+                error.Errors[0].Code);
+            user = await this.dbContext.Users.FindAsync(user.Id);
+            Assert.Equal("Alice2", user.UserName);
+            Assert.Equal(
+                new byte[10], user.PasswordHash);
+
+            this.dbContext.Remove(user);
+            await this.dbContext.SaveChangesAsync();
+        }
+
+        [Fact]
+        public async Task TestRemove()
+        {
+            var user = new User(
+                "Alice2", new byte[10], NullGenerator.Identifier, "{}");
+            this.dbContext.Add(user);
+            await this.dbContext.SaveChangesAsync();
+
+            await this.userManager.Remove(user.Id);
+
+            var users = this.dbContext.Users.Where(u => u.UserName == "Alice2");
+            Assert.Empty(users);
+        }
+
+        [Fact]
+        public async Task TestRemoveWrongId()
+        {
+            var error = await Assert.ThrowsAsync<QueryException>(async () =>
+            {
+                await this.userManager.Remove(-5);
+            });
+
+            Assert.Equal(
+                Error.UserNotFound.ToString(),
+                error.Errors[0].Code);
         }
     }
 }