+++ CHANGELOG.md
@@ -5,7 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Query endpoint "Users" to view and filter available users.
+- Mutation endpoint "CreateUser" to add a user to the application.
+- Mutation endpoint "UpdateUser" to update a user of the application.
+- Mutation endpoint "RemoveUser" to remove a user from the application.
+- Mutation endpoint "CreateDeploymentPlan" to replace "AddDeploymentPlan"
 
+### Removed
+- Mutation endpoint "AddDeploymentPlan" for consistent naming.
 
 ## [0.2.0] - 2021-05-18
 ### Added